requirements.txt
Django==1.9.2
django-angular==0.7.16
django-tastypie==0.13.3
freeze==1.0.9
python-dateutil==2.4.2
python-mimeparse==1.5.1
six==1.10.0
wheel==0.29.0
